remove order logic in optimizer, move usd to common folder
