+++ src/common/CMakeLists.txt
@@ -67,7 +67,9 @@ add_library_unity(
   tree_renderer.cpp
   types.cpp
   virtual_file_system.cpp
-  windows_util.cpp)
+  windows_util.cpp
+  unified_string_dictionary.cpp
+)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_common>
     PARENT_SCOPE)
