+++ src/execution/operator/helper/physical_unified_string_dictionary_insertion.cpp
@@ -22,16 +22,16 @@ class USDInsertionGState : public GlobalOperatorState {
 public:
 	explicit USDInsertionGState(ClientContext &context, idx_t cols) {
 		for (idx_t i = 0; i < cols; ++i) {
-			inserted_unique_strings.push_back(0);
-			unique_strings_in_unified_dictionary_per_column.push_back(0);
+			inserted_strings.push_back(0);
+			unique_strings_in_usd_per_column.push_back(0);
 			inserted_dictionaries.push_back(0);
 			is_high_cardinality.push_back(false);
 		}
 	}
 	mutex statistics_lock;
 	vector<bool> is_high_cardinality;
-	vector<idx_t> inserted_unique_strings;
-	vector<idx_t> unique_strings_in_unified_dictionary_per_column;
+	vector<idx_t> inserted_strings;
+	vector<idx_t> unique_strings_in_usd_per_column;
 	vector<idx_t> inserted_dictionaries;
 };
 
@@ -40,6 +40,7 @@ OperatorResultType PhysicalUnifiedStringDictionary::Execute(ExecutionContext &co
                                                             OperatorState &state_p) const {
 	auto &state = state_p.Cast<USDInsertionState>();
 	auto &global_state = gstate.Cast<USDInsertionGState>();
+
 	for (idx_t col_idx = 0; col_idx < input.data.size(); ++col_idx) {
 		if (input.data[col_idx].GetType() != LogicalType::VARCHAR || !insert_to_usd[col_idx]) {
 			continue;
@@ -95,8 +96,8 @@ OperatorResultType PhysicalUnifiedStringDictionary::Execute(ExecutionContext &co
 				// update local and global states
 				state.current_dict_ids[col_idx] = DictionaryVector::DictionaryId(input.data[col_idx]);
 				unique_lock<mutex> lock(global_state.statistics_lock);
-				global_state.inserted_unique_strings[col_idx] += size.GetIndex();
-				global_state.unique_strings_in_unified_dictionary_per_column[col_idx] += state.n_success;
+				global_state.inserted_strings[col_idx] += size.GetIndex();
+				global_state.unique_strings_in_usd_per_column[col_idx] += state.n_success;
 				global_state.inserted_dictionaries[col_idx]++;
 				// FIXME: magic numbers should not be here, still trying to fine tune this section
 				constexpr double TOTAL_GROWTH_THRESHOLD = 0.1;
@@ -105,21 +106,20 @@ OperatorResultType PhysicalUnifiedStringDictionary::Execute(ExecutionContext &co
 
 				if (global_state.inserted_dictionaries[col_idx] > MIN_DICTIONARY_SEEN) {
 					auto avg_growth =
-					    static_cast<double>(global_state.unique_strings_in_unified_dictionary_per_column[col_idx]) /
-					    static_cast<double>(global_state.inserted_unique_strings[col_idx]);
+					    static_cast<double>(global_state.unique_strings_in_usd_per_column[col_idx]) /
+					    static_cast<double>(global_state.inserted_strings[col_idx]);
 
 					if (avg_growth > TOTAL_GROWTH_THRESHOLD) {
 						global_state.is_high_cardinality[col_idx] = true;
 					}
 				}
-				if (global_state.unique_strings_in_unified_dictionary_per_column[col_idx] > HARD_LIMIT) {
+				if (global_state.unique_strings_in_usd_per_column[col_idx] > HARD_LIMIT) {
 					global_state.is_high_cardinality[col_idx] = true;
 				}
 				lock.unlock();
 
 				context.client.GetUnifiedStringDictionary().UpdateFailedAttempts(state.n_rejected_probing +
 				                                                                 state.n_rejected_full);
-				// FIXME: move the whole statistics update out of execution
 				state.n_success = 0;
 				state.n_rejected_full = 0;
 				state.n_rejected_probing = 0;
