+++ src/include/duckdb/common/unified_string_dictionary.h
@@ -11,10 +11,6 @@
 
 namespace duckdb {
 
-#if !defined(DUCKDB_DISABLE_POINTER_SALT) && defined(__ANDROID__)
-#define DUCKDB_DISABLE_POINTER_SALT
-#endif
-
 enum class InsertResult {
 	// string is inserted for the first time
 	SUCCESS,
@@ -30,10 +26,10 @@ enum class InsertResult {
 
 // Unified String Dictionary is a per-query dictionary containing the most valuable strings in a particular query with
 // their pre-computed hashes. The strings and their hashes are materialized in the Data Region. A linear probing hash
-// table is used for fast look ups.
+// table is used for fast look up.
 class UnifiedStringsDictionary {
 public:
-	UnifiedStringsDictionary(idx_t usd_sf);
+	explicit UnifiedStringsDictionary(idx_t usd_sf);
 	~UnifiedStringsDictionary();
 	void UpdateFailedAttempts(idx_t n_failed);
 	InsertResult Insert(string_t &str);
