+++ src/include/duckdb/main/query_result.hpp
@@ -13,7 +13,7 @@
 #include "duckdb/common/winapi.hpp"
 #include "duckdb/common/error_data.hpp"
 #include "duckdb/main/client_properties.hpp"
-#include "duckdb/optimizer/unified_string_dictionary.h"
+#include "duckdb/common/unified_string_dictionary.h"
 
 namespace duckdb {
 struct BoxRendererConfig;
