+++ src/optimizer/CMakeLists.txt
@@ -37,7 +37,6 @@ add_library_unity(
   unnest_rewriter.cpp
   sampling_pushdown.cpp
   sum_rewriter.cpp
-  unified_string_dictionary.cpp
   unified_string_dictionary_optimizer.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_optimizer>
